Fix time check before sleeping
Previously, a simple limit of sleepDuration > 5 seconds was required to enter sleep. In the event of a prolonged measurement cycle (from slow sensor response or SD card shuffling, the alarm we set at the beginning of the measurement could have already passed. Now, the limit is if there is at least 5 seconds between now and the alarm time.